+++ phpmyfaq/lang/language_de.php
@@ -1338,10 +1338,12 @@ $PMF_LANG['ad_add_user_change_password'] = 'Der Nutzer muss nach dem ersten Logi
 $LANG_CONF['api.enableAccess'] = ['checkbox', 'REST API aktiviert'];
 $LANG_CONF['api.apiClientToken'] = ['input', 'REST API Client Token'];
 $LANG_CONF['security.domainWhiteListForRegistrations'] = ['area', 'Erlaubte Domains bei Registrierungen'];
-$LANG_CONF['main.loginWithEmailAddress'] = ['checkbox', 'Login nur mit E-Mailadresse'];
+$LANG_CONF['security.loginWithEmailAddress'] = ['checkbox', 'Login nur mit E-Mailadresse'];
 
 // added v3.2.0-alpha - 2022-09-10 by Thorsten
 $PMF_LANG['msgSignInWithMicrosoft'] = 'Mit Microsoft anmelden';
-$LANG_CONF['security.enableSignInWithMicrosoft'] = ['checkbox', 'Aktiviere Anmeldung mit Microsoft'];
+$LANG_CONF['security.enableSignInWithMicrosoft'] = ['checkbox', 'Aktiviere Anmeldung mit Microsoft (Azure AD)'];
+$LANG_CONF['main.enableAskQuestions'] = ['checkbox', 'Aktiviere "Frage stellen"'];
+$LANG_CONF['main.enableNotifications'] = ['checkbox', 'Aktiviere Benachrichtigungen'];
 
 return $PMF_LANG;