+++ phpmyfaq/lang/language_ja.php
@@ -1077,6 +1077,6 @@ $PMF_LANG['ad_add_user_change_password'] = '初回ログイン時にはパスワ
 $LANG_CONF['api.enableAccess'] = ['checkbox', 'REST API有効化'];
 $LANG_CONF['api.apiClientToken'] = ['input', 'API Client Token'];
 $LANG_CONF['security.domainWhiteListForRegistrations'] = ['area', 'ユーザー登録の許可ホワイトリスト'];
-$LANG_CONF['main.loginWithEmailAddress'] = ['checkbox', 'メールアドレスによるログインのみ許可'];
+$LANG_CONF['security.loginWithEmailAddress'] = ['checkbox', 'メールアドレスによるログインのみ許可'];
 
 return $PMF_LANG;