+++ phpmyfaq/lang/language_pt-br.php
@@ -1050,7 +1050,7 @@ $PMF_LANG['ad_add_user_change_password'] = 'O usuário deve alterar a senha apó
 $LANG_CONF['api.enableAccess'] = ["checkbox", "Habilitar REST API"];
 $LANG_CONF['api.apiClientToken'] = ["input", "Token do client da API"];
 $LANG_CONF['security.domainWhiteListForRegistrations'] = ["area", "Hosts permitidos para cadastros"];
-$LANG_CONF['main.loginWithEmailAddress'] = ["checkbox", "Login somente com endereço de email"];
+$LANG_CONF['security.loginWithEmailAddress'] = ["checkbox", "Login somente com endereço de email"];
 $PMF_LANG["plmsgUserOnline"][1] = '%d usuários online';
 
 return $PMF_LANG;